Nowa klasa
Nowa klasa - szykowanie pod problem wielomaszynowy
