+++ Labolatorium 1/__main__.py
@@ -8,22 +8,28 @@ if __name__ == "__main__":
     tasks = [Task(i) for i in range(1,6)]
     A = Machine(1)
     B = Machine(2)
+    C = Machine(3)
+    D = Machine(4)
 
     time_A = [4, 4, 10, 6, 2]
     time_B = [5, 1, 4, 10, 3]
+    time_C = [4, 2, 6, 1, 5]
+
     for i in range(len(tasks)):
         task = tasks[i]
         A.add_task_duration(task, time_A[i])
         B.add_task_duration(task, time_B[i])
+        C.add_task_duration(task, time_C[i])
+        D.add_task_duration(task, time_B[i])
 
     # algorytm
-    algorithm = SillyAlgorithm()
-    machines_with_task = algorithm.run([A, B], tasks)
+    algorithm = JohnsonRule()
+    machines_with_task = algorithm.run([A, B, C D], tasks)
 
     # wyswietlenie i obliczenia
     for machine in machines_with_task:
         print(machine)
 
     gantt = Gantt(machines_with_task)
-    print(f"Czas trwania{gantt.get_duration()}")
+    print(f"Czas trwania: {gantt.get_duration()}")
     gantt.plot()
